+++ libraries/google/dataprocmetastore/property/service_metadata_integration_dataplex_config_lake_resources.rb
@@ -13,27 +13,20 @@
 #     CONTRIBUTING.md located at the root of this package.
 #
 # ----------------------------------------------------------------------------
-require 'google/dataprocmetastore/property/projectlocationservice_scaling_config_autoscaling_config_limit_config'
 module GoogleInSpec
   module DataprocMetastore
     module Property
-      class ProjectLocationServiceScalingConfigAutoscalingConfig
-        attr_reader :autoscaling_factor
-
-        attr_reader :autoscaling_enabled
-
-        attr_reader :limit_config
+      class ServiceMetadataIntegrationDataplexConfigLakeResources
+        attr_reader :additional_properties
 
         def initialize(args = nil, parent_identifier = nil)
           return if args.nil?
           @parent_identifier = parent_identifier
-          @autoscaling_factor = args['autoscalingFactor']
-          @autoscaling_enabled = args['autoscalingEnabled']
-          @limit_config = GoogleInSpec::DataprocMetastore::Property::ProjectLocationServiceScalingConfigAutoscalingConfigLimitConfig.new(args['limitConfig'], to_s)
+          @additional_properties = args['additionalProperties']
         end
 
         def to_s
-          "#{@parent_identifier} ProjectLocationServiceScalingConfigAutoscalingConfig"
+          "#{@parent_identifier} ServiceMetadataIntegrationDataplexConfigLakeResources"
         end
       end
     end
