+++ libraries/google/dataprocmetastore/property/service_metadata_management_activity_restores.rb
@@ -16,37 +16,43 @@
 module GoogleInSpec
   module DataprocMetastore
     module Property
-      class ProjectLocationServiceMetadataManagementActivityMetadataExports
-        attr_reader :destination_gcs_uri
-
+      class ServiceMetadataManagementActivityRestores
         attr_reader :start_time
 
         attr_reader :end_time
 
         attr_reader :state
 
-        attr_reader :database_dump_type
+        attr_reader :backup
+
+        attr_reader :type
+
+        attr_reader :details
+
+        attr_reader :backup_location
 
         def initialize(args = nil, parent_identifier = nil)
           return if args.nil?
           @parent_identifier = parent_identifier
-          @destination_gcs_uri = args['destinationGcsUri']
           @start_time = args['startTime']
           @end_time = args['endTime']
           @state = args['state']
-          @database_dump_type = args['databaseDumpType']
+          @backup = args['backup']
+          @type = args['type']
+          @details = args['details']
+          @backup_location = args['backupLocation']
         end
 
         def to_s
-          "#{@parent_identifier} ProjectLocationServiceMetadataManagementActivityMetadataExports"
+          "#{@parent_identifier} ServiceMetadataManagementActivityRestores"
         end
       end
 
-      class ProjectLocationServiceMetadataManagementActivityMetadataExportsArray
+      class ServiceMetadataManagementActivityRestoresArray
         def self.parse(value, parent_identifier)
           return if value.nil?
-          return ProjectLocationServiceMetadataManagementActivityMetadataExports.new(value, parent_identifier) unless value.is_a?(::Array)
-          value.map { |v| ProjectLocationServiceMetadataManagementActivityMetadataExports.new(v, parent_identifier) }
+          return ServiceMetadataManagementActivityRestores.new(value, parent_identifier) unless value.is_a?(::Array)
+          value.map { |v| ServiceMetadataManagementActivityRestores.new(v, parent_identifier) }
         end
       end
     end
