+++ test/integration/verify/controls/google_dataproc_metastore_service.rb
@@ -12,7 +12,7 @@
 #
 # ----------------------------------------------------------------------------
 
-title 'Test GCP google_dataproc_metastore_project_location_service resource.'
+title 'Test GCP google_dataproc_metastore_service resource.'
 
 gcp_project_id = input(:gcp_project_id, value: 'gcp_project_id', description: 'The GCP project identifier.')
 
@@ -31,11 +31,11 @@ gcp_project_id = input(:gcp_project_id, value: 'gcp_project_id', description: 'T
   "release_channel": "value_releasechannel",
   "database_type": "value_databasetype"
 }, description: 'project_location_service description')
-control 'google_dataproc_metastore_project_location_service-1.0' do
+control 'google_dataproc_metastore_service-1.0' do
   impact 1.0
-  title 'google_dataproc_metastore_project_location_service resource test'
+  title 'google_dataproc_metastore_service resource test'
 
-  describe google_dataproc_metastore_project_location_service(name: project_location_service['name']) do
+  describe google_dataproc_metastore_service(name: project_location_service['name']) do
   	it { should exist }
   	its('name') { should cmp project_location_service['name'] }
   	its('create_time') { should cmp project_location_service['create_time'] }
@@ -52,7 +52,7 @@ control 'google_dataproc_metastore_project_location_service-1.0' do
 
   end
 
-  describe google_dataproc_metastore_project_location_service(name: "does_not_exit") do
+  describe google_dataproc_metastore_service(name: "does_not_exit") do
   	it { should_not exist }
   end
 end
